requirements.txt
tensorflow
numpy
pillow
tensorflow-hub